requirements.txt
Django
pytz
celery
redis
psycopg2
djangorestframework
dj-database-url
graypy
raven
uwsgi

